requirements.txt
# fitforge-api/requirements.txt
together==0.11.0
fastapi==0.115.0
together==0.11.0
uvicorn[standard]==0.34.0
together==0.11.0
supabase==2.0.0
together==0.11.0
groq==0.13.0
together==0.11.0
google-generativeai==0.8.0
together==0.11.0
httpx==0.27.0
together==0.11.0
python-jose[cryptography]==3.3.0
together==0.11.0
python-dotenv==1.0.1
together==0.11.0
pydantic==2.10.0
together==0.11.0
pydantic-settings==2.6.0
together==0.11.0
